Module Simple_httpd

Http Server

This library implements a very simple, basic HTTP/1.1 server using blocking IOs and threads. Basic routing based on Scanf is provided for convenience, so that several handlers can be registered.

It is possible to use a thread pool, see create's argument new_thread.

The echo example (see src/examples/echo.ml) demonstrates some of the features by declaring a few endpoints, including one for uploading files:

module S = Simple_httpd

let () =
  let server = S.create () in

  (* say hello *)
  S.add_route_handler ~meth:`GET server
    S.Route.(exact "hello" @/ string @/ return)
    (fun name _req -> S.Response.make_string (Ok ("hello " ^name ^"!\n")));

  (* echo request *)
  S.add_route_handler server
    S.Route.(exact "echo" @/ return)
    (fun req -> S.Response.make_string
        (Ok (Format.asprintf "echo:@ %a@." S.Request.pp req)));

  (* file upload *)
  S.add_route_handler ~meth:`PUT server
    S.Route.(exact "upload" @/ string_urlencoded @/ return)
    (fun path req ->
        try
          let oc = open_out @@ "/tmp/" ^ path in
          output_string oc req.S.Request.body;
          flush oc;
          S.Response.make_string (Ok "uploaded file")
        with e ->
          S.Response.fail ~code:500 "couldn't upload file: %s"
            (Printexc.to_string e)
      );

  (* run the server *)
  Printf.printf "listening on http://%s:%d\n%!" (S.addr server) (S.port server);
  match S.run server with
  | Ok () -> ()
  | Error e -> raise e

It is then possible to query it using curl:

$ dune exec src/examples/echo.exe &
listening on http://127.0.0.1:8080

# the path "hello/name" greets you.
$ curl -X GET http://localhost:8080/hello/quadrarotaphile
hello quadrarotaphile!

# the path "echo" just prints the request.
$ curl -X GET http://localhost:8080/echo --data "howdy y'all"
echo:
{meth=GET;
 headers=Host: localhost:8080
         User-Agent: curl/7.66.0
         Accept: */*
         Content-Length: 10
         Content-Type: application/x-www-form-urlencoded;
 path="/echo"; body="howdy y'all"}

Tiny buffer implementation

These buffers are used to avoid allocating too many byte arrays when processing streams and parsing requests.

module Buf = Simple_httpd_buf

Generic stream of data

Streams are used to represent a series of bytes that can arrive progressively. For example, an uploaded file will be sent as a series of chunks.

module Byte_stream = Simple_httpd_stream

Main Server Type

type buf = Simple_httpd_buf.t

type byte_stream = Simple_httpd_stream.t

Methods

module Meth = Simple_httpd_server.Meth

module Cookies = Simple_httpd_server.Cookies

Headers

Headers are metadata associated with a request or response.

module Headers = Simple_httpd_server.Headers

Requests

Requests are sent by a client, e.g. a web browser or cURL.

module Request = Simple_httpd_server.Request

Response Codes

module Response_code = Simple_httpd_server.Response_code

Responses

Responses are what a http server, such as Simple_httpd, send back to the client to answer a Request.t

module Response = Simple_httpd_server.Response

module Route = Simple_httpd_server.Route

module Middleware = Simple_httpd_server.Middleware

Main Server type

type t = Simple_httpd_server.t

A HTTP server. See create for more details.

val create : ?masksigpipe:bool -> ?max_connections:int -> ?granularity:int -> ?num_thread:int -> ?timeout:float -> ?buf_size:int -> ?get_time_s:(unit -> float) -> ?listens:Simple_httpd_domain.listenning list -> ?middlewares:([ `Encoding | `Stage of int ] * Middleware.t) list -> unit -> t

Create a new webserver.

The server will not do anything until run is called on it. Before starting the server, one can use add_path_handler and set_top_handler to specify how to handle incoming requests.

parameter masksigpipe
if true, block the signal Sys.sigpipe which otherwise tends to kill client threads when they try to write on broken sockets. Default: true.

parameter buf_size
size for buffers (since 0.11)

parameter middlewares
see add_middleware for more details.

parameter max_connections
maximum number of simultaneous connections.

parameter granularity
yield every granularity read or write.

parameter num_thread
number of thread to treat client.

parameter timeout
connection is closed if the socket does not do read or write for the amount of second. Default: 0.0 which means no timeout. timeout is not recommended when using proxy.

parameter addr
address (IPv4 or IPv6) to listen on. Default "127.0.0.1".

parameter port
to listen on. Default 8080.

parameter sock
an existing socket given to the server to listen on, e.g. by systemd on Linux (or launchd on macOS). If passed in, this socket will be used instead of the addr and port. If not passed in, those will be used. This parameter exists since 0.10.

parameter get_time_s
obtain the current timestamp in seconds. This parameter exists since 0.11.

val listens : t -> Simple_httpd_domain.listenning list

Addresses and ports on which the server listens.

val active_connections : t -> int

Number of active connections

val add_decode_request_cb : t -> (unit Request.t -> (unit Request.t * (byte_stream -> byte_stream)) option) -> unit

Add a callback for every request. The callback can provide a stream transformer and a new request (with modified headers, typically). A possible use is to handle decompression by looking for a Transfer-Encoding header and returning a stream transformer that decompresses on the fly.

deprecated
use add_middleware instead

val add_encode_response_cb : t -> (unit Request.t -> Response.t -> Response.t option) -> unit

Add a callback for every request/response pair. Similarly to add_encode_response_cb the callback can return a new response, for example to compress it. The callback is given the query with only its headers, as well as the current response.

deprecated
use add_middleware instead

val add_middleware : stage:[ `Encoding | `Stage of int ] -> t -> Middleware.t -> unit

Add a middleware to every request/response pair.

parameter stage
specify when middleware applies. Encoding comes first (outermost layer), then stages in increasing order.

raises Invalid_argument
if stage is `Stage n where n < 1

Request handlers

val set_top_handler : t -> (string Request.t -> Response.t) -> unit

Setup a handler called by default.

This handler is called with any request not accepted by any handler installed via add_path_handler. If no top handler is installed, unhandled paths will return a 404 not found.

type finaliser = Response.t -> Response.t

type 'a accept = 'a Request.t -> (finaliser, Response_code.t * string) Stdlib.result

val add_route_handler : ?accept:unit accept -> ?middlewares:Middleware.t list -> ?meth:Meth.t -> t -> ('a, string Request.t -> Response.t) Route.t -> 'a -> unit

add_route_handler server Route.(exact "path" @/ string @/ int @/ return) f calls f "foo" 42 request when a request with path "path/foo/42/" is received.

Note that the handlers are called in the reverse order of their addition, so the last registered handler can override previously registered ones.

parameter meth
if provided, only accept requests with the given method. Typically one could react to `GET or `PUT.

parameter accept
should return Ok fn if the given request (before its body is read) should be accepted, Error (code,message) if it's to be rejected (e.g. because. fn: finaliser is used to transform the response. its content is too big, or for some permission error). See the http_of_dir program for an example of how to use accept to filter uploads that are too large before the upload even starts. The default always returns Ok(), i.e. it accepts all requests.

val add_route_handler_stream : ?accept:unit accept -> ?middlewares:Middleware.t list -> ?meth:Meth.t -> t -> ('a, byte_stream Request.t -> Response.t) Route.t -> 'a -> unit

Similar to add_route_handler, but where the body of the request is a stream of bytes that has not been read yet. This is useful when one wants to stream the body directly into a parser, json decoder (such as Jsonm) or into a file.

Server-sent events

EXPERIMENTAL: this API is not stable yet.

module type SERVER_SENT_GENERATOR = Simple_httpd_server.SERVER_SENT_GENERATOR

A server-side function to generate of Server-sent events.

type server_sent_generator = (module SERVER_SENT_GENERATOR)

Server-sent event generator

val add_route_server_sent_handler : ?accept:unit accept -> t -> ('a, string Request.t -> server_sent_generator -> unit) Route.t -> 'a -> unit

Add a handler on an endpoint, that serves server-sent events.

The callback is given a generator that can be used to send events as it pleases. The connection is always closed by the client, and the accepted method is always GET. This will set the header "content-type" to "text/event-stream" automatically and reply with a 200 immediately. See server_sent_generator for more details.

This handler stays on the original thread (it is synchronous).

Run the server

val run : t -> (unit, exn) Stdlib.result

Run the main loop of the server, listening on a socket described at the server's creation time, using new_thread to start a thread for each new client.

This returns Ok () if the server exits gracefully, or Error e if it exits with an error.

Utils

module Util = Simple_httpd_util

Static directory serving

module Dir = Simple_httpd_dir

module Html = Simple_httpd_html

Alias to Simple_httpd_html

module Domain = Simple_httpd_domain

module Session = Simple_httpd_session

type listenning = Simple_httpd_domain.listenning = {

addr : string;

port : int;

ssl : Ssl.context option;

}

usefull type

cooperative threading

The following functions deals with cooperative multi-tasking on each domain. First, recall that the OS will choose via select on which domain you serve your client, and after that, it is impossible for a job to change domain. This is a current limitation of domain with OCaml 5.0.

Then, on each domain, priority is based on arrival time: first arrived, first to run. With one exception: Mutexes have the highest priority, and among mutexes, the highest priority is again, first arrived, first to run.

If you want to have mutex with lower priority, you may do: yield (); lock m or if not (Mutex.try_lock m) then (yield (); lock m)

This should more or less give a priority for mutexes comparable to other scheduled jobs... With more risk of deadlock ?

Normally context switching occurs when read or write is blocked in the socket serving the client or regularly (based on granularity parameter)

val yield : unit -> unit

let other thread run. Should be called for treatment that take time before sending results or reading data and when the other primitive can not be used. This happends typically for a pure computing task. Other solutions exists for instance for database request.

val sleep : float -> unit

Same as above, but with a minimum sleeping time in second

exception Closed of bool

exception used by the two functions below

module Io : sig ... end

Module with function similar to Unix.read and Unix.single_write but that will perform scheduling

val schedule_read : Unix.file_descr -> (unit -> int) -> (exn -> unit) -> int

schedule_read sock action close should be called when a non blocking read operation would have blocked. When read become possible, action () will be called. If it raises an exception exn, close exn will be called it is up to the user to decide of closing the socket and/or reraise the exception.

The return value should be (if possible) the number of bytes read. It this is meaningless, return a non zero value if some progress was made, while returning 0 will call close (Closed true).

A typical application for this is when interacting with a data base in non blocking mode. For just reading a file or socket, use the Io module above.

val schedule_write : Unix.file_descr -> (unit -> int) -> (exn -> unit) -> int

schedule_read sock action close is similar as above for a write operation. If it return action () return 0; close (Closed false) is called.

val lock : Stdlib.Mutex.t -> unit