Module Simple_httpd_server.Cookies

Set Cookie

A module to set new cookies in the header

type t

val empty : t

val parse : string -> t

val add : string -> Http_cookie.t -> t -> t

val create : ?path:string -> ?domain:string -> ?expires:Http_cookie.date_time -> ?max_age:int64 -> ?secure:bool -> ?http_only:bool -> ?same_site:Http_cookie.same_site -> ?extension:string -> name:string -> string -> t -> t

val get : string -> t -> Http_cookie.t

val delete : string -> t -> t

remove a cookie by setting a negative max-age. Does nothing if there are no cookie with that name.

val delete_all : t -> t

remove all cookies by setting a negative max-age